table changes and refactoring
