+++ crates/nu-table/.vscode/launch.json
@@ -1,0 +1,64 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "type": "lldb",
+            "request": "launch",
+            "name": "Debug unit tests in library 'nu-table'",
+            "cargo": {
+                "args": [
+                    "test",
+                    "--no-run",
+                    "--lib",
+                    "--package=nu-table"
+                ],
+                "filter": {
+                    "name": "nu-table",
+                    "kind": "lib"
+                }
+            },
+            "args": [],
+            "cwd": "${workspaceFolder}"
+        },
+        {
+            "type": "lldb",
+            "request": "launch",
+            "name": "Debug executable 'table'",
+            "cargo": {
+                "args": [
+                    "build",
+                    "--bin=table",
+                    "--package=nu-table"
+                ],
+                "filter": {
+                    "name": "table",
+                    "kind": "bin"
+                }
+            },
+            "args": ["120"],
+            "cwd": "${workspaceFolder}"
+        },
+        {
+            "type": "lldb",
+            "request": "launch",
+            "name": "Debug unit tests in executable 'table'",
+            "cargo": {
+                "args": [
+                    "test",
+                    "--no-run",
+                    "--bin=table",
+                    "--package=nu-table"
+                ],
+                "filter": {
+                    "name": "table",
+                    "kind": "bin"
+                }
+            },
+            "args": [],
+            "cwd": "${workspaceFolder}"
+        },
+    ]
+}
