+++ crates/nu-table/src/lib.rs
@@ -1,5 +1,8 @@
+pub mod styled_string;
 mod table;
+pub mod table_theme;
+pub mod text_style;
 mod wrap;
 
-pub use table::{draw_table, StyledString, Table, TextStyle, Theme};
+pub use table::{draw_table, Table};
 pub use wrap::Alignment;
