+++ crates/nu-table/src/table.rs
@@ -1,5 +1,10 @@
-use crate::wrap::{column_width, split_sublines, wrap, Alignment, Subline, WrappedCell};
-use ansi_term::{Color, Style};
+use crate::styled_string::StyledString;
+use crate::table_theme::TableTheme;
+use crate::text_style::TextStyle;
+use crate::wrap::{
+    column_width, split_sublines, wrap, Alignment, Subline, WrappedCell, WrappedLine,
+};
+use ansi_term::Style;
 use std::collections::HashMap;
 
 enum SeparatorPosition {
@@ -12,591 +17,15 @@ enum SeparatorPosition {
 pub struct Table {
     pub headers: Vec<StyledString>,
     pub data: Vec<Vec<StyledString>>,
-    pub theme: Theme,
-}
-
-#[derive(Debug, Clone)]
-pub struct StyledString {
-    pub contents: String,
-    pub style: TextStyle,
-}
-
-impl StyledString {
-    pub fn new(contents: String, style: TextStyle) -> StyledString {
-        StyledString { contents, style }
-    }
-
-    pub fn set_style(&mut self, style: TextStyle) {
-        self.style = style;
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-pub struct TextStyle {
-    pub alignment: Alignment,
-    pub color_style: Option<Style>,
-}
-
-impl TextStyle {
-    pub fn new() -> TextStyle {
-        TextStyle {
-            alignment: Alignment::Left,
-            color_style: Some(Style::default()),
-        }
-    }
-
-    pub fn bold(&self, bool_value: Option<bool>) -> TextStyle {
-        let bv = bool_value.unwrap_or(false);
-
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                is_bold: bv,
-                ..self.color_style.unwrap_or_default()
-            }),
-        }
-    }
-
-    pub fn is_bold(&self) -> bool {
-        self.color_style.unwrap_or_default().is_bold
-    }
-
-    pub fn dimmed(&self) -> TextStyle {
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                is_dimmed: true,
-                ..self.color_style.unwrap_or_default()
-            }),
-        }
-    }
-
-    pub fn is_dimmed(&self) -> bool {
-        self.color_style.unwrap_or_default().is_dimmed
-    }
-
-    pub fn italic(&self) -> TextStyle {
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                is_italic: true,
-                ..self.color_style.unwrap_or_default()
-            }),
-        }
-    }
-
-    pub fn is_italic(&self) -> bool {
-        self.color_style.unwrap_or_default().is_italic
-    }
-
-    pub fn underline(&self) -> TextStyle {
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                is_underline: true,
-                ..self.color_style.unwrap_or_default()
-            }),
-        }
-    }
-
-    pub fn is_underline(&self) -> bool {
-        self.color_style.unwrap_or_default().is_underline
-    }
-
-    pub fn blink(&self) -> TextStyle {
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                is_blink: true,
-                ..self.color_style.unwrap_or_default()
-            }),
-        }
-    }
-
-    pub fn is_blink(&self) -> bool {
-        self.color_style.unwrap_or_default().is_blink
-    }
-
-    pub fn reverse(&self) -> TextStyle {
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                is_reverse: true,
-                ..self.color_style.unwrap_or_default()
-            }),
-        }
-    }
-
-    pub fn is_reverse(&self) -> bool {
-        self.color_style.unwrap_or_default().is_reverse
-    }
-
-    pub fn hidden(&self) -> TextStyle {
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                is_hidden: true,
-                ..self.color_style.unwrap_or_default()
-            }),
-        }
-    }
-
-    pub fn is_hidden(&self) -> bool {
-        self.color_style.unwrap_or_default().is_hidden
-    }
-
-    pub fn strikethrough(&self) -> TextStyle {
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                is_strikethrough: true,
-                ..self.color_style.unwrap_or_default()
-            }),
-        }
-    }
-
-    pub fn is_strikethrough(&self) -> bool {
-        self.color_style.unwrap_or_default().is_strikethrough
-    }
-
-    pub fn fg(&self, foreground: Color) -> TextStyle {
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                foreground: Some(foreground),
-                ..self.color_style.unwrap_or_default()
-            }),
-        }
-    }
-
-    pub fn on(&self, background: Color) -> TextStyle {
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                background: Some(background),
-                ..self.color_style.unwrap_or_default()
-            }),
-        }
-    }
-
-    pub fn bg(&self, background: Color) -> TextStyle {
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                background: Some(background),
-                ..self.color_style.unwrap_or_default()
-            }),
-        }
-    }
-
-    pub fn alignment(&self, align: Alignment) -> TextStyle {
-        TextStyle {
-            alignment: align,
-            color_style: self.color_style,
-        }
-    }
-
-    pub fn style(&self, style: Style) -> TextStyle {
-        TextStyle {
-            alignment: self.alignment,
-            color_style: Some(Style {
-                foreground: style.foreground,
-                background: style.background,
-                is_bold: style.is_bold,
-                is_dimmed: style.is_dimmed,
-                is_italic: style.is_italic,
-                is_underline: style.is_underline,
-                is_blink: style.is_blink,
-                is_reverse: style.is_reverse,
-                is_hidden: style.is_hidden,
-                is_strikethrough: style.is_strikethrough,
-            }),
-        }
-    }
-
-    pub fn basic_center() -> TextStyle {
-        TextStyle::new()
-            .alignment(Alignment::Center)
-            .style(Style::default())
-    }
-
-    pub fn basic_right() -> TextStyle {
-        TextStyle::new()
-            .alignment(Alignment::Right)
-            .style(Style::default())
-    }
-
-    pub fn basic_left() -> TextStyle {
-        TextStyle::new()
-            .alignment(Alignment::Left)
-            .style(Style::default())
-    }
-
-    pub fn default_header() -> TextStyle {
-        TextStyle::new()
-            .alignment(Alignment::Center)
-            .fg(Color::Green)
-            .bold(Some(true))
-    }
-
-    pub fn with_attributes(bo: bool, al: Alignment, co: Color) -> TextStyle {
-        TextStyle::new().alignment(al).fg(co).bold(Some(bo))
-    }
-
-    pub fn with_style(al: Alignment, style: Style) -> TextStyle {
-        TextStyle::new().alignment(al).style(Style {
-            foreground: style.foreground,
-            background: style.background,
-            is_bold: style.is_bold,
-            is_dimmed: style.is_dimmed,
-            is_italic: style.is_italic,
-            is_underline: style.is_underline,
-            is_blink: style.is_blink,
-            is_reverse: style.is_reverse,
-            is_hidden: style.is_hidden,
-            is_strikethrough: style.is_strikethrough,
-        })
-    }
-}
-
-impl Default for TextStyle {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct Theme {
-    pub top_left: char,
-    pub middle_left: char,
-    pub bottom_left: char,
-    pub top_center: char,
-    pub center: char,
-    pub bottom_center: char,
-    pub top_right: char,
-    pub middle_right: char,
-    pub bottom_right: char,
-    pub top_horizontal: char,
-    pub middle_horizontal: char,
-    pub bottom_horizontal: char,
-    pub left_vertical: char,
-    pub center_vertical: char,
-    pub right_vertical: char,
-
-    pub separate_header: bool,
-    pub separate_rows: bool,
-
-    pub print_left_border: bool,
-    pub print_right_border: bool,
-    pub print_top_border: bool,
-    pub print_bottom_border: bool,
-}
-
-impl Theme {
-    #[allow(unused)]
-    pub fn basic() -> Theme {
-        Theme {
-            top_left: '+',
-            middle_left: '+',
-            bottom_left: '+',
-            top_center: '+',
-            center: '+',
-            bottom_center: '+',
-            top_right: '+',
-            middle_right: '+',
-            bottom_right: '+',
-            top_horizontal: '-',
-            middle_horizontal: '-',
-            bottom_horizontal: '-',
-            left_vertical: '|',
-            center_vertical: '|',
-            right_vertical: '|',
-
-            separate_header: true,
-            separate_rows: true,
-
-            print_left_border: true,
-            print_right_border: true,
-            print_top_border: true,
-            print_bottom_border: true,
-        }
-    }
-
-    #[allow(unused)]
-    pub fn thin() -> Theme {
-        Theme {
-            top_left: '┌',
-            middle_left: '├',
-            bottom_left: '└',
-            top_center: '┬',
-            center: '┼',
-            bottom_center: '┴',
-            top_right: '┐',
-            middle_right: '┤',
-            bottom_right: '┘',
-
-            top_horizontal: '─',
-            middle_horizontal: '─',
-            bottom_horizontal: '─',
-
-            left_vertical: '│',
-            center_vertical: '│',
-            right_vertical: '│',
-
-            separate_header: true,
-            separate_rows: true,
-
-            print_left_border: true,
-            print_right_border: true,
-            print_top_border: true,
-            print_bottom_border: true,
-        }
-    }
-
-    #[allow(unused)]
-    pub fn light() -> Theme {
-        Theme {
-            top_left: ' ',
-            middle_left: '─',
-            bottom_left: ' ',
-            top_center: ' ',
-            center: '─',
-            bottom_center: ' ',
-            top_right: ' ',
-            middle_right: '─',
-            bottom_right: ' ',
-
-            top_horizontal: ' ',
-            middle_horizontal: '─',
-            bottom_horizontal: ' ',
-
-            left_vertical: ' ',
-            center_vertical: ' ',
-            right_vertical: ' ',
-
-            separate_header: true,
-            separate_rows: false,
-
-            print_left_border: true,
-            print_right_border: true,
-            print_top_border: false,
-            print_bottom_border: true,
-        }
-    }
-
-    #[allow(unused)]
-    pub fn compact() -> Theme {
-        Theme {
-            top_left: '─',
-            middle_left: '─',
-            bottom_left: '─',
-            top_center: '┬',
-            center: '┼',
-            bottom_center: '┴',
-            top_right: '─',
-            middle_right: '─',
-            bottom_right: '─',
-            top_horizontal: '─',
-            middle_horizontal: '─',
-            bottom_horizontal: '─',
-
-            left_vertical: ' ',
-            center_vertical: '│',
-            right_vertical: ' ',
-
-            separate_header: true,
-            separate_rows: false,
-
-            print_left_border: false,
-            print_right_border: false,
-            print_top_border: true,
-            print_bottom_border: true,
-        }
-    }
-
-    #[allow(unused)]
-    pub fn with_love() -> Theme {
-        Theme {
-            top_left: '❤',
-            middle_left: '❤',
-            bottom_left: '❤',
-            top_center: '❤',
-            center: '❤',
-            bottom_center: '❤',
-            top_right: '❤',
-            middle_right: '❤',
-            bottom_right: '❤',
-            top_horizontal: '❤',
-            middle_horizontal: '❤',
-            bottom_horizontal: '❤',
-
-            left_vertical: ' ',
-            center_vertical: '❤',
-            right_vertical: ' ',
-
-            separate_header: true,
-            separate_rows: false,
-
-            print_left_border: false,
-            print_right_border: false,
-            print_top_border: true,
-            print_bottom_border: true,
-        }
-    }
-
-    #[allow(unused)]
-    pub fn compact_double() -> Theme {
-        Theme {
-            top_left: '═',
-            middle_left: '═',
-            bottom_left: '═',
-            top_center: '╦',
-            center: '╬',
-            bottom_center: '╩',
-            top_right: '═',
-            middle_right: '═',
-            bottom_right: '═',
-            top_horizontal: '═',
-            middle_horizontal: '═',
-            bottom_horizontal: '═',
-
-            left_vertical: ' ',
-            center_vertical: '║',
-            right_vertical: ' ',
-
-            separate_header: true,
-            separate_rows: false,
-
-            print_left_border: false,
-            print_right_border: false,
-            print_top_border: true,
-            print_bottom_border: true,
-        }
-    }
-
-    #[allow(unused)]
-    pub fn rounded() -> Theme {
-        Theme {
-            top_left: '╭',
-            middle_left: '├',
-            bottom_left: '╰',
-            top_center: '┬',
-            center: '┼',
-            bottom_center: '┴',
-            top_right: '╮',
-            middle_right: '┤',
-            bottom_right: '╯',
-            top_horizontal: '─',
-            middle_horizontal: '─',
-            bottom_horizontal: '─',
-
-            left_vertical: '│',
-            center_vertical: '│',
-            right_vertical: '│',
-
-            separate_header: true,
-            separate_rows: false,
-
-            print_left_border: true,
-            print_right_border: true,
-            print_top_border: true,
-            print_bottom_border: true,
-        }
-    }
-
-    #[allow(unused)]
-    pub fn reinforced() -> Theme {
-        Theme {
-            top_left: '┏',
-            middle_left: '├',
-            bottom_left: '┗',
-            top_center: '┬',
-            center: '┼',
-            bottom_center: '┴',
-            top_right: '┓',
-            middle_right: '┤',
-            bottom_right: '┛',
-            top_horizontal: '─',
-            middle_horizontal: '─',
-            bottom_horizontal: '─',
-
-            left_vertical: '│',
-            center_vertical: '│',
-            right_vertical: '│',
-
-            separate_header: true,
-            separate_rows: false,
-
-            print_left_border: true,
-            print_right_border: true,
-            print_top_border: true,
-            print_bottom_border: true,
-        }
-    }
-
-    #[allow(unused)]
-    pub fn heavy() -> Theme {
-        Theme {
-            top_left: '┏',
-            middle_left: '┣',
-            bottom_left: '┗',
-            top_center: '┳',
-            center: '╋',
-            bottom_center: '┻',
-            top_right: '┓',
-            middle_right: '┫',
-            bottom_right: '┛',
-            top_horizontal: '━',
-            middle_horizontal: '━',
-            bottom_horizontal: '━',
-
-            left_vertical: '┃',
-            center_vertical: '┃',
-            right_vertical: '┃',
-
-            separate_header: true,
-            separate_rows: false,
-
-            print_left_border: true,
-            print_right_border: true,
-            print_top_border: true,
-            print_bottom_border: true,
-        }
-    }
-    #[allow(unused)]
-    pub fn none() -> Theme {
-        Theme {
-            top_left: ' ',
-            middle_left: ' ',
-            bottom_left: ' ',
-            top_center: ' ',
-            center: ' ',
-            bottom_center: ' ',
-            top_right: ' ',
-            middle_right: ' ',
-            bottom_right: ' ',
-
-            top_horizontal: ' ',
-            middle_horizontal: ' ',
-            bottom_horizontal: ' ',
-
-            left_vertical: ' ',
-            center_vertical: ' ',
-            right_vertical: ' ',
-
-            separate_header: false,
-            separate_rows: false,
-
-            print_left_border: false,
-            print_right_border: false,
-            print_top_border: false,
-            print_bottom_border: false,
-        }
-    }
+    pub theme: TableTheme,
 }
 
 impl Table {
-    pub fn new(headers: Vec<StyledString>, data: Vec<Vec<StyledString>>, theme: Theme) -> Table {
+    pub fn new(
+        headers: Vec<StyledString>,
+        data: Vec<Vec<StyledString>>,
+        theme: TableTheme,
+    ) -> Table {
         Table {
             headers,
             data,
@@ -609,7 +38,7 @@ impl Table {
 pub struct ProcessedTable<'a> {
     pub headers: Vec<ProcessedCell<'a>>,
     pub data: Vec<Vec<ProcessedCell<'a>>>,
-    pub theme: Theme,
+    pub theme: TableTheme,
 }
 
 #[derive(Debug)]
@@ -623,7 +52,7 @@ pub struct WrappedTable {
     pub column_widths: Vec<usize>,
     pub headers: Vec<WrappedCell>,
     pub data: Vec<Vec<WrappedCell>>,
-    pub theme: Theme,
+    pub theme: TableTheme,
 }
 
 impl WrappedTable {
@@ -963,104 +392,229 @@ fn get_max_column_widths(processed_table: &ProcessedTable) -> Vec<usize> {
     output
 }
 
-pub fn maybe_truncate_columns(termwidth: usize, processed_table: &mut ProcessedTable) {
-    // Make sure we have enough space for the columns we have
-    let max_num_of_columns = termwidth / 10;
+// pub fn maybe_truncate_columns(
+//     termwidth: usize,
+//     processed_table: &mut ProcessedTable,
+//     max_column_widths: &[usize],
+// ) {
+//     let (_running_column_width, running_column_count) =
+//         get_column_fit_in_termwidth(termwidth, max_column_widths);
 
-    // If we have too many columns, truncate the table
-    if max_num_of_columns < processed_table.headers.len() {
-        processed_table.headers.truncate(max_num_of_columns);
+//     // Make sure we have enough space for the columns we have
+//     let max_num_of_columns = running_column_count - 2;
+//     // eprintln!("max_num_of_columns={}", max_num_of_columns);
+//     // If we have too many columns, truncate the table
+//     if max_num_of_columns < processed_table.headers.len() {
+//         processed_table.headers.truncate(max_num_of_columns);
 
-        for entry in processed_table.data.iter_mut() {
-            entry.truncate(max_num_of_columns);
-        }
+//         for entry in processed_table.data.iter_mut() {
+//             entry.truncate(max_num_of_columns);
+//         }
 
-        processed_table.headers.push(ProcessedCell {
-            contents: vec![vec![Subline {
-                subline: "...",
-                width: 3,
-            }]],
-            style: TextStyle::basic_center(),
-        });
+//         processed_table.headers.push(ProcessedCell {
+//             contents: vec![vec![Subline {
+//                 subline: "...",
+//                 width: 3,
+//             }]],
+//             style: TextStyle::basic_center(),
+//         });
 
-        for entry in processed_table.data.iter_mut() {
-            entry.push(ProcessedCell {
-                contents: vec![vec![Subline {
-                    subline: "...",
-                    width: 3,
-                }]],
-                style: TextStyle::basic_center(),
-            }); // ellipsis is centred
-        }
-    }
-}
+//         for entry in processed_table.data.iter_mut() {
+//             entry.push(ProcessedCell {
+//                 contents: vec![vec![Subline {
+//                     subline: "...",
+//                     width: 3,
+//                 }]],
+//                 style: TextStyle::basic_center(),
+//             }); // ellipsis is centred
+//         }
+//     }
+// }
 
 pub fn draw_table(table: &Table, termwidth: usize, color_hm: &HashMap<String, Style>) {
     // Remove the edges, if used
-    let termwidth = if table.theme.print_left_border && table.theme.print_right_border {
-        termwidth - 2
-    } else if table.theme.print_left_border || table.theme.print_right_border {
-        termwidth - 1
-    } else {
-        termwidth
-    };
+    // let termwidth = if table.theme.print_left_border && table.theme.print_right_border {
+    //     termwidth - 2
+    // } else if table.theme.print_left_border || table.theme.print_right_border {
+    //     termwidth - 1
+    // } else {
+    //     termwidth
+    // };
 
-    let mut processed_table = process_table(table);
+    let processed_table = process_table(table);
 
+    // get the max column width for each column using the header and rows
     let max_per_column = get_max_column_widths(&processed_table);
 
-    maybe_truncate_columns(termwidth, &mut processed_table);
+    // maybe_truncate_columns(termwidth, &mut processed_table, &max_per_column);
 
-    let headers_len = processed_table.headers.len();
+    // let headers_len = processed_table.headers.len();
 
-    // fix the length of the table if there are no headers:
-    let headers_len = if headers_len == 0 {
-        if !table.data.is_empty() && !table.data[0].is_empty() {
-            table.data[0].len()
-        } else {
-            return;
-        }
-    } else {
-        headers_len
-    };
+    // // fix the length of the table if there are no headers:
+    // let headers_len = if headers_len == 0 {
+    //     if !table.data.is_empty() && !table.data[0].is_empty() {
+    //         table.data[0].len()
+    //     } else {
+    //         return;
+    //     }
+    // } else {
+    //     headers_len
+    // };
 
     // Measure how big our columns need to be (accounting for separators also)
-    let max_naive_column_width = (termwidth - 3 * (headers_len - 1)) / headers_len;
+    // let max_naive_column_width = (termwidth - 2 * (headers_len - 1)) / headers_len;
 
-    let column_space = ColumnSpace::measure(&max_per_column, max_naive_column_width, headers_len);
+    // let column_space = ColumnSpace::measure(&max_per_column, max_naive_column_width, headers_len);
 
     // This gives us the max column width
-    let max_column_width = column_space.max_width(termwidth);
+    // let max_column_width = column_space.max_width(termwidth);
 
     // This width isn't quite right, as we're rounding off some of our space
-    let column_space = column_space.fix_almost_column_width(
-        &max_per_column,
-        max_naive_column_width,
-        max_column_width,
-        headers_len,
-    );
+    // let column_space = column_space.fix_almost_column_width(
+    //     &max_per_column,
+    //     max_naive_column_width,
+    //     max_column_width,
+    //     headers_len,
+    // );
 
     // This should give us the final max column width
-    let max_column_width = column_space.max_width(termwidth);
+    // let max_column_width = column_space.max_width(termwidth);
     let re_leading =
         regex::Regex::new(r"(?P<beginsp>^\s+)").expect("error with leading space regex");
     let re_trailing =
         regex::Regex::new(r"(?P<endsp>\s+$)").expect("error with trailing space regex");
 
-    let wrapped_table = wrap_cells(
+    let (_total_width, total_columns) = get_column_fit_in_termwidth(termwidth, &max_per_column);
+
+    // We really only need this info for the columns we're going to try and fit
+
+    // Get the widest column out of the collection
+    let mut modified_max_per_column = max_per_column;
+    // eprintln!(
+    //     "termwidth=[{}] max_per_col=[{}] total_columns=[{}]",
+    //     termwidth,
+    //     modified_max_per_column.len(),
+    //     total_columns
+    // );
+    if modified_max_per_column[0] == 1 && modified_max_per_column.len() > total_columns {
+        modified_max_per_column = modified_max_per_column[1..=total_columns].to_vec()
+    } else {
+        modified_max_per_column.truncate(total_columns);
+    }
+    // eprint!(" mod_max_per_col=[{:?}]", modified_max_per_column);
+    let max_column_width = *modified_max_per_column
+        .iter()
+        .max()
+        .expect("failed getting max of columns");
+    // eprint!(" max_col_wid=[{}]", max_column_width);
+    // Add up the widths of every column
+    let all_columns_width_sum: usize = modified_max_per_column.iter().sum();
+    // eprint!(" all_col_wid_sum=[{}]", all_columns_width_sum);
+    let avg_column_width = all_columns_width_sum / total_columns;
+    // eprint!(" avg_col_wid=[{}]", avg_column_width);
+    // If the terminal isn't wide enough to fit all columns, modify it
+    let mut _modified_max_column_width = 0usize;
+    if all_columns_width_sum > termwidth {
+        _modified_max_column_width = avg_column_width;
+    } else {
+        _modified_max_column_width = max_column_width;
+    }
+    // eprintln!(" mod_max_col_wid=[{}]", _modified_max_column_width);
+    // maybe_truncate_columns(termwidth, &mut processed_table, &modified_max_per_column);
+
+    let mut wrapped_table = wrap_cells(
         processed_table,
-        max_column_width,
+        _modified_max_column_width,
+        total_columns,
         &color_hm,
         &re_leading,
         &re_trailing,
     );
+    // eprintln!("mod_max_per_col=[{:?}]", &modified_max_per_column);
+    maybe_truncate_wrapped_columns(&mut wrapped_table, termwidth, &modified_max_per_column);
 
     wrapped_table.print_table(&color_hm);
+}
+
+fn get_column_fit_in_termwidth(termwidth: usize, max_column_widths: &[usize]) -> (usize, usize) {
+    // How many of the real columns will fit in this max width
+    // the number of lines between columns not including the far left and far right
+    let column_separators: usize = max_column_widths.len() - 1;
+    // start at 2 for far left and far right table line
+    let column_left_and_right = 2usize;
+    // start point is column separators plus far left and right table lines
+    let mut running_column_width = column_separators + column_left_and_right;
+    let mut running_column_count = 0usize;
+    let mut last_width = 0usize;
+    for current_width in max_column_widths.iter() {
+        if running_column_width > termwidth {
+            running_column_width -= last_width + 2;
+            running_column_count -= 1;
+            break;
+        } else {
+            running_column_width += current_width + 2; // each column has at least 1 space on each side
+            running_column_count += 1;
+            last_width = *current_width;
+        }
+    }
+
+    (running_column_width, running_column_count)
+}
+
+fn maybe_truncate_wrapped_columns(
+    wrapped_table: &mut WrappedTable,
+    termwidth: usize,
+    max_column_widths: &[usize],
+) {
+    let (_running_column_width, running_column_count) =
+        get_column_fit_in_termwidth(termwidth, max_column_widths);
+
+    // Make sure we have enough space for the columns we have
+    let max_num_of_columns = running_column_count;
+
+    if max_num_of_columns < wrapped_table.headers.len() {
+        // Chop off the headers that are too wide for our terminal
+        wrapped_table.headers.truncate(max_num_of_columns);
+
+        // Chop off the data that are too wide for our terminal
+        for entry in wrapped_table.data.iter_mut() {
+            entry.truncate(max_num_of_columns);
+        }
+
+        // Chop off the column_widths too
+        wrapped_table.column_widths.truncate(max_num_of_columns);
+
+        // Insert ... for header
+        wrapped_table.headers.push(WrappedCell {
+            lines: vec![WrappedLine {
+                line: "...".to_string(),
+                width: 3,
+            }],
+            max_width: 3,
+            style: TextStyle::basic_center(),
+        });
+
+        // Insert ... for data
+        for entry in wrapped_table.data.iter_mut() {
+            entry.push(WrappedCell {
+                lines: vec![WrappedLine {
+                    line: "...".to_string(),
+                    width: 3,
+                }],
+                max_width: 3,
+                style: TextStyle::basic_center(),
+            });
+        }
+
+        // Insert a column width for ...
+        wrapped_table.column_widths.push(3);
+    }
 }
 
 fn wrap_cells(
     processed_table: ProcessedTable,
     max_column_width: usize,
+    total_columns: usize,
     color_hm: &HashMap<String, Style>,
     re_leading: &regex::Regex,
     re_trailing: &regex::Regex,
@@ -1076,8 +630,19 @@ fn wrap_cells(
             }
         )
     ];
+
+    // Color for leading and trailing space color from config.toml
+    let lead_trail_space_bg_color = color_hm
+        .get("leading_trailing_space_bg")
+        .unwrap_or(&Style::default())
+        .to_owned();
+
     let mut output_headers = vec![];
     for header in processed_table.headers.into_iter().enumerate() {
+        // only process what will show up on screen
+        if header.0 > total_columns {
+            break;
+        }
         let mut wrapped = WrappedCell {
             lines: vec![],
             max_width: 0,
@@ -1088,9 +653,9 @@ fn wrap_cells(
             let (mut lines, inner_max_width) = wrap(
                 max_column_width,
                 contents.into_iter(),
-                &color_hm,
                 &re_leading,
                 &re_trailing,
+                &lead_trail_space_bg_color,
             );
             wrapped.lines.append(&mut lines);
             if inner_max_width > wrapped.max_width {
@@ -1107,6 +672,9 @@ fn wrap_cells(
     for row in processed_table.data.into_iter() {
         let mut output_row = vec![];
         for column in row.into_iter().enumerate() {
+            if column.0 > total_columns {
+                break;
+            }
             let mut wrapped = WrappedCell {
                 lines: vec![],
                 max_width: 0,
@@ -1116,9 +684,9 @@ fn wrap_cells(
                 let (mut lines, inner_max_width) = wrap(
                     max_column_width,
                     contents.into_iter(),
-                    &color_hm,
                     &re_leading,
                     &re_trailing,
+                    &lead_trail_space_bg_color,
                 );
                 wrapped.lines.append(&mut lines);
                 if inner_max_width > wrapped.max_width {
@@ -1141,106 +709,106 @@ fn wrap_cells(
     }
 }
 
-struct ColumnSpace {
-    num_overages: usize,
-    underage_sum: usize,
-    overage_separator_sum: usize,
-}
+// struct ColumnSpace {
+//     num_overages: usize,
+//     underage_sum: usize,
+//     overage_separator_sum: usize,
+// }
 
-impl ColumnSpace {
-    /// Measure how much space we have once we subtract off the columns who are small enough
-    fn measure(
-        max_per_column: &[usize],
-        max_naive_column_width: usize,
-        headers_len: usize,
-    ) -> ColumnSpace {
-        let mut num_overages = 0;
-        let mut underage_sum = 0;
-        let mut overage_separator_sum = 0;
-        let iter = max_per_column.iter().enumerate().take(headers_len);
+// impl ColumnSpace {
+//     /// Measure how much space we have once we subtract off the columns who are small enough
+//     fn measure(
+//         max_per_column: &[usize],
+//         max_naive_column_width: usize,
+//         headers_len: usize,
+//     ) -> ColumnSpace {
+//         let mut num_overages = 0;
+//         let mut underage_sum = 0;
+//         let mut overage_separator_sum = 0;
+//         let iter = max_per_column.iter().enumerate().take(headers_len);
 
-        for (i, &column_max) in iter {
-            if column_max > max_naive_column_width {
-                num_overages += 1;
-                if i != (headers_len - 1) {
-                    overage_separator_sum += 3;
-                }
-                if i == 0 {
-                    overage_separator_sum += 1;
-                }
-            } else {
-                underage_sum += column_max;
-                // if column isn't last, add 3 for its separator
-                if i != (headers_len - 1) {
-                    underage_sum += 3;
-                }
-                if i == 0 {
-                    underage_sum += 1;
-                }
-            }
-        }
+//         for (i, &column_max) in iter {
+//             if column_max > max_naive_column_width {
+//                 num_overages += 1;
+//                 if i != (headers_len - 1) {
+//                     overage_separator_sum += 3;
+//                 }
+//                 if i == 0 {
+//                     overage_separator_sum += 1;
+//                 }
+//             } else {
+//                 underage_sum += column_max;
+//                 // if column isn't last, add 3 for its separator
+//                 if i != (headers_len - 1) {
+//                     underage_sum += 3;
+//                 }
+//                 if i == 0 {
+//                     underage_sum += 1;
+//                 }
+//             }
+//         }
 
-        ColumnSpace {
-            num_overages,
-            underage_sum,
-            overage_separator_sum,
-        }
-    }
+//         ColumnSpace {
+//             num_overages,
+//             underage_sum,
+//             overage_separator_sum,
+//         }
+//     }
 
-    fn fix_almost_column_width(
-        self,
-        max_per_column: &[usize],
-        max_naive_column_width: usize,
-        max_column_width: usize,
-        headers_len: usize,
-    ) -> ColumnSpace {
-        let mut num_overages = 0;
-        let mut overage_separator_sum = 0;
-        let mut underage_sum = self.underage_sum;
-        let iter = max_per_column.iter().enumerate().take(headers_len);
+//     fn fix_almost_column_width(
+//         self,
+//         max_per_column: &[usize],
+//         max_naive_column_width: usize,
+//         max_column_width: usize,
+//         headers_len: usize,
+//     ) -> ColumnSpace {
+//         let mut num_overages = 0;
+//         let mut overage_separator_sum = 0;
+//         let mut underage_sum = self.underage_sum;
+//         let iter = max_per_column.iter().enumerate().take(headers_len);
 
-        for (i, &column_max) in iter {
-            if column_max > max_naive_column_width {
-                if column_max <= max_column_width {
-                    underage_sum += column_max;
-                    // if column isn't last, add 3 for its separator
-                    if i != (headers_len - 1) {
-                        underage_sum += 3;
-                    }
-                    if i == 0 {
-                        underage_sum += 1;
-                    }
-                } else {
-                    // Column is still too large, so let's count it
-                    num_overages += 1;
-                    if i != (headers_len - 1) {
-                        overage_separator_sum += 3;
-                    }
-                    if i == 0 {
-                        overage_separator_sum += 1;
-                    }
-                }
-            }
-        }
+//         for (i, &column_max) in iter {
+//             if column_max > max_naive_column_width {
+//                 if column_max <= max_column_width {
+//                     underage_sum += column_max;
+//                     // if column isn't last, add 3 for its separator
+//                     if i != (headers_len - 1) {
+//                         underage_sum += 3;
+//                     }
+//                     if i == 0 {
+//                         underage_sum += 1;
+//                     }
+//                 } else {
+//                     // Column is still too large, so let's count it
+//                     num_overages += 1;
+//                     if i != (headers_len - 1) {
+//                         overage_separator_sum += 3;
+//                     }
+//                     if i == 0 {
+//                         overage_separator_sum += 1;
+//                     }
+//                 }
+//             }
+//         }
 
-        ColumnSpace {
-            num_overages,
-            underage_sum,
-            overage_separator_sum,
-        }
-    }
+//         ColumnSpace {
+//             num_overages,
+//             underage_sum,
+//             overage_separator_sum,
+//         }
+//     }
 
-    fn max_width(&self, termwidth: usize) -> usize {
-        let ColumnSpace {
-            num_overages,
-            underage_sum,
-            overage_separator_sum,
-        } = self;
+//     fn max_width(&self, termwidth: usize) -> usize {
+//         let ColumnSpace {
+//             num_overages,
+//             underage_sum,
+//             overage_separator_sum,
+//         } = self;
 
-        if *num_overages > 0 {
-            (termwidth - 1 - *underage_sum - *overage_separator_sum) / *num_overages
-        } else {
-            99999
-        }
-    }
-}
+//         if *num_overages > 0 {
+//             (termwidth - 1 - *underage_sum - *overage_separator_sum) / *num_overages
+//         } else {
+//             99999
+//         }
+//     }
+// }
